+++ tests/TextStack.AiEvals/CapturingDb.cs
@@ -98,5 +98,6 @@ internal sealed class CapturingDb : IAppDbContext
     public DbSet<AgentRun> AgentRuns => throw new NotSupportedException();
     public DbSet<ShadowRun> ShadowRuns => throw new NotSupportedException();
     public DbSet<ModelRegistration> Models => throw new NotSupportedException();
+    public DbSet<ModelPromotion> ModelPromotions => throw new NotSupportedException();
     public DbSet<PodcastGenerationJob> PodcastGenerationJobs => throw new NotSupportedException();
 }